rename: constant 패키지 명 수정
constant -> message
